+++ backend/database/migrations/2025_10_26_100550_contao_round_publish_flags.php
@@ -16,7 +16,7 @@ return new class extends Migration
             try {
                 // Create table first without foreign key
                 Schema::create('contao_public_rounds', function (Blueprint $table) {
-                    $table->unsignedBigInteger('event_id')->primary();
+                    $table->unsignedInteger('event_id')->primary();
                     $table->boolean('vr1')->default(true);
                     $table->boolean('vr2')->default(false);
                     $table->boolean('vr3')->default(false);
